+++ src/alchemlyb/preprocessing/__init__.py
@@ -12,11 +12,3 @@ __all__ = [
     'statistical_inefficiency',
     'equilibrium_detection',
 ]
-
-def pass_attrs(func):
-    '''Pass the attributes from the input dataframe to the output dataframe'''
-    def wrapper(input_dataframe, *args,**kwargs):
-        dataframe = func(input_dataframe, *args,**kwargs)
-        dataframe.attrs = input_dataframe.attrs
-        return dataframe
-    return wrapper
